+++ src/money/index.spec.ts
@@ -1,4 +1,4 @@
-import { Dollar } from '../money';
+import Money, { Dollar } from '../money';
 import Franc from '../franc';
 
 describe('Money', () => {
@@ -9,6 +9,26 @@ describe('Money', () => {
       expect(new Dollar(5).equals(new Dollar(5))).toBe(true);
       expect(new Dollar(5).equals(new Dollar(6))).toBe(false);
       expect(new Franc(5).equals(new Dollar(5))).toBe(false);
+    });
+  });
+});
+
+describe('Dollar', () => {
+  it.todo('should sum $5 + 10CHF as being $10 if exchange rate is 2:1');
+
+  describe('multiplication', () => {
+    it('should multiply $5 by 2 and get $10', () => {
+      const five: Money = Money.dollar(5);
+      const product = five.times(2);
+      expect(product.equals(Money.dollar(10))).toBe(true);
+    });
+
+    it('should preserve the initial dollar amount when multiplying multiple times', () => {
+      const five = Money.dollar(5);
+      let product: Dollar = five.times(2);
+      expect(product.equals(Money.dollar(10))).toBe(true);
+      product = five.times(3);
+      expect(product.equals(Money.dollar(15))).toBe(true);
     });
   });
 });
